+++ main.hpp
@@ -53,6 +53,8 @@ extern size_t max_tile_features;
 extern int cluster_distance;
 extern std::string attribute_for_id;
 extern int tiny_polygon_size;
+extern size_t limit_tile_feature_count;
+extern size_t limit_tile_feature_count_at_maxzoom;
 
 struct order_field {
 	std::string name;