+++ mbtiles.cpp
@@ -372,6 +372,12 @@ std::string stringify_strategies(std::vector<strategy> const &strategies) {
 			any = true;
 		}
 
+		if (strategies[i].feature_count > 0) {
+			state.json_write_string("feature_count_desired");
+			state.json_write_number(strategies[i].feature_count);
+			any = true;
+		}
+
 		state.json_end_hash();
 	}
 	state.json_end_array();